+++ platform/data/staticProvider_test.go
@@ -178,9 +178,7 @@ func TestStaticProvider_AddDataToContext(t *testing.T) {
 
 		newCtx, err := provider.AddDataToContext(
 			ctx,
-			map[string]any{"key": "value"},
-			map[string]any{"str": "string"},
-			map[string]any{"num": 42},
+			map[string]any{"key": "value", "str": "string", "num": 42},
 		)
 
 		require.Error(t, err, "StaticProvider should reject all attempts to add data")